Refactor: move functionality to new files
